Rename project related to domains UI and reorganize packages
