+++ bundles/extensions/emfdomain/tools.vitruv.extensions.emf/META-INF/MANIFEST.MF
@@ -13,7 +13,7 @@ Require-Bundle: org.eclipse.core.resources,
  org.eclipse.core.expressions,
  org.apache.log4j,
  org.eclipse.emf.ecore,
- tools.vitruv.framework.ui.monitorededitor;visibility:=reexport,
+ tools.vitruv.framework.domains.ui;visibility:=reexport,
  org.eclipse.emf.transaction,
  tools.vitruv.framework.vsum,
  tools.vitruv.extensions.emf.monitorededitor,
