+++ bundles/framework/tools.vitruv.framework.domains.ui/.classpath
