+++ bundles/framework/tools.vitruv.framework.domains.ui/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>tools.vitruv.framework.ui.monitorededitor</name>
+	<name>tools.vitruv.framework.domains.ui</name>
 	<comment></comment>
 	<projects>
 	</projects>
