+++ bundles/framework/tools.vitruv.framework.domains.ui/build.properties
