+++ bundles/framework/tools.vitruv.framework.domains.ui/src/tools/vitruv/framework/domains/ui/Activator.java
@@ -1,4 +1,4 @@
-package tools.vitruv.framework.ui.monitorededitor;
+package tools.vitruv.framework.domains.ui;
 
 import org.osgi.framework.BundleActivator;
 import org.osgi.framework.BundleContext;
