+++ bundles/framework/tools.vitruv.framework.domains.ui/src/tools/vitruv/framework/domains/ui/builder/TestBuilderEnabled.java
@@ -1,4 +1,4 @@
-package tools.vitruv.framework.ui.monitorededitor;
+package tools.vitruv.framework.domains.ui.builder;
 
 import org.eclipse.core.expressions.PropertyTester;
 import org.eclipse.core.resources.IProject;
