+++ bundles/framework/tools.vitruv.framework.domains.ui/src/tools/vitruv/framework/domains/ui/builder/VitruvProjectBuilderArguments.xtend
@@ -1,4 +1,4 @@
-package tools.vitruv.framework.ui.monitorededitor
+package tools.vitruv.framework.domains.ui.builder
 
 import org.eclipse.core.resources.ICommand
 import java.io.File
