+++ bundles/framework/tools.vitruv.framework.domains.ui/src/tools/vitruv/framework/ui/monitorededitor/registries/package-info.java
