+++ bundles/framework/tools.vitruv.framework.ui.vsum/META-INF/MANIFEST.MF
@@ -15,7 +15,7 @@ Require-Bundle: org.eclipse.ui,
  org.eclipse.xtend.lib,
  org.eclipse.xtend.lib.macro,
  tools.vitruv.framework.vsum,
- tools.vitruv.framework.ui.monitorededitor,
+ tools.vitruv.framework.domains.ui,
  org.apache.log4j
 Bundle-RequiredExecutionEnvironment: JavaSE-11
 Bundle-ActivationPolicy: lazy
